Fix prefix calculation for non-ascii ccache records
If the C implementation uses (signed char) for (char), then non-ascii
paths may have a shorter suffix than they should.  Same rationale as:
    2025-03-07 Explicit casts for char to int
    6072b2ce897531eb2e32fdb531657730a9eb09df

Reported by:	clang-tidy-15 bugprone-signed-char-misuse

+++ tar/ccache/ccache_write.c
@@ -113,7 +113,7 @@ callback_write_rec(void * cookie, uint8_t * s, size_t slen, void * rec)
 
 	/* Figure out how much prefix is shared. */
 	for (plen = 0; plen < slen && plen < W->sbuflen; plen++) {
-		if (s[plen] != W->sbuf[plen])
+		if (s[plen] != (uint8_t)W->sbuf[plen])
 			break;
 	}
